Optimize month parsing and tests.
Do not use objects but `constexpr' C arrays and strings, to create
simple in-memory tables.
Do not create a temporary object with the all lower case representation
of the input string, instead convert each character to lower case when
it is compared to the reference string.
Use a separate object module for the `main()' function of Catch2, which
takes long to compile.