Fix build on FreeBSD
On FreeBSD, the libc implementation also has an `_assert' macro. This
collides with the `_assert' function in the `wildland_firesim:utility'
namespace. Renaming it to `verboseAssert' resolves this issue while also
being a bit more specific on the functions intention.

Also, the `clang++' 7.0 on FreeBSD complains about the seed argument of
the random number generator being narrowed down, so instead of using a
generic `int' the type is now dependent on the `result_type' of the
random number generator.